+++ apps/frontend/awesome/src/main.tsx
@@ -5,7 +5,7 @@ import App from './App.tsx';
 import './index.css';
 
 createRoot(document.getElementById('root')!).render(
-  // <StrictMode>
-  <App />,
-  // </StrictMode>,
+  <StrictMode>
+    <App />
+  </StrictMode>,
 );
